Updated GHUnit to support arm64. Added some typecasting for ints in order to GHUnit tests to pass (type mismatching errors with test constant literals).

+++ OCSquirrelTests/TestBindClass.m
@@ -476,7 +476,7 @@
     
     InitializerChecker *object = instance.instanceUP;
     
-    XCTAssertEqual(object.intProperty, 12345,
+    XCTAssertEqual((NSInteger)object.intProperty, (NSInteger)12345,
                    @"Initializer method should exist and accept the int parameter.");
 }
